http/sign: add Signature.HasRightTime()
HasRightTime checks if the signature was issued at the right time.
The signature is considered to be issued at the right time if the difference
between the signature time and the specified time is less than the specified precision.
The signature with zero time is always considered to be issued at the right time.

Change-Id: I3e15ac914c446b39cda3bf47f8781dba381254d1