[stacktrace] introduce new type Frame
Frame represents a single stack frame containing program counter, file
path, line number, and function name. This is a simplified
representation of runtime.Frame that is safe to store and use after the
stack trace has been captured.

Dedicated structure is safer for passing around and allows for
independent testing, not as a part of Stack.
Also this structure will be used in future method that will capture the
caller.

Change-Id: I648a0ad8f07dcfc491dc9cdd1fe552a24e6fe06e